requirements.txt
# Core dependencies for wallpaper collection system
# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# HTTP requests and web scraping
requests>=2.31.0
urllib3>=2.0.0

# Image quality assessment (requires opencv-contrib-python)
opencv-contrib-python>=4.8.0

# JSON and data handling
jsonschema>=4.0.0

# Logging and utilities
colorlog>=6.7.0
tqdm>=4.65.0

# Date and time handling
python-dateutil>=2.8.0

# Optional: For advanced features
# beautifulsoup4>=4.12.0  # For web scraping
# selenium>=4.15.0  # For dynamic content scraping
# psutil>=5.9.0  # For system resource monitoring

# Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Development tools (optional)
# flake8>=6.0.0
# black>=23.0.0
# mypy>=1.5.0